https://jira.jboss.org/browse/JBIDE-7377 Exception in error log after several files deletion -> ResourceException: Resource 'xxx' does not exist. Fixed.
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@25965 a97e2381-89e5-4abb-bab3-167db6db766c

+++ plugins/org.jboss.tools.jst.web.kb/src/org/jboss/tools/jst/web/kb/internal/validation/ValidatorManager.java
@@ -101,7 +101,9 @@ public class ValidatorManager implements IValidatorJob {
 	private void removeMarkers(Set<IFile> files) {
 		try {
 			for (IFile file : files) {
-				file.deleteMarkers(IValidator.KB_PROBLEM_MARKER_TYPE, true, IResource.DEPTH_ZERO);
+				if(file.isAccessible()) {
+					file.deleteMarkers(IValidator.KB_PROBLEM_MARKER_TYPE, true, IResource.DEPTH_ZERO);
+				}
 			}
 		} catch (CoreException e) {
 			WebKbPlugin.getDefault().logError(e);
